+++ cmd/rampart/cli/serve_test.go
@@ -380,7 +380,7 @@ func TestServeStatePublishesPrivateExecutableIdentity(t *testing.T) {
 		t.Fatal(err)
 	}
 	const pid = 4242
-	if err := writeServeState(dir, 19090, pid, false, proxy.RuntimeIdentity{}); err != nil {
+	if err := writeServeState(dir, 19090, pid, false, proxy.RuntimeIdentity{}, nil); err != nil {
 		t.Fatal(err)
 	}
 	want, err := os.Executable()